refactor(polar_alignment): split pipeline stages, move MC benchmark out
Behaviour-preserving restructure; the unit tests are untouched and the
seeded Monte Carlo benchmark reproduces the previous table exactly.

- extract _epoch_align_solves, _two_solve_axis and _ignore_roll_axis
  from get_platform_adjustments so the main function reads as the
  pipeline its numbered comments describe
- extract _altaz_knob_rotation from correction_target (also the seam
  where the annual-aberration pole nudge plugs back in)
- move the ~200-line Monte Carlo benchmark to
  PiFinder/polar_alignment_benchmark.py (python -m PiFinder.polar_alignment_benchmark)
- reuse _boresight_vec for the axis vector in correction_target, drop
  the unused _axis_from_dec_ra and the benchmark-only logging setup
